I moved code around and tested everything with the Program.cs test file. I'll need to change its name later. The test program creates a brain and adds one input and one output neuron. I got the results that I expected by changeing the inputs over a range and firing them. I still need to test a larger array of neurons and include hidden neurons.
I also added functionality to attach a function to the neurons that runs when they fire. Currently, it only runs for output neurons. I anticipate this action will be whatever cellular function the output neurons control. It is currently set to always fire, which completely underminest the purpose of the network. I will need to add a threshhold so that the function is only run if the activation of the output is above this value. I have reminded myself of this in a TODO.